retet lidt mere double

+++ Repetition Inheritance/BenzinBil.cs
@@ -16,7 +16,7 @@ namespace Repetition_Inheritance
         public int Tank { get; set; }
         public override double HalvÅrligEjerafgift()
         {
-            int ejerafgft = 0;
+            double ejerafgft = 0;
             if (KmPrLiter < 20)
             {
                 ejerafgft = 1200;
@@ -34,7 +34,7 @@ namespace Repetition_Inheritance
 
         public override double Rækkevidde()
         {
-            int række = Tank*KmPrLiter;
+            double række = Tank*KmPrLiter;
             return række;
         }
     }
